chore: allow wider `engine` version range
This is not a NodeJS package but a browser package, thus we do not care
about the package manager - its only a development dependency. So move
it to `devEngines`.
As we support apps for all supported Nextcloud versions, we need to
support current Node (22) and also the Node version of the stable
Nextcloud version (Node 20).
And we should be future proof and support current active Node 24.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>